Final commit

+++ ProjetoCAL/src/graphviewer.cpp
@@ -93,7 +93,7 @@ bool GraphViewer::closeWindow() {
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::addNode(unsigned long id) {
+bool GraphViewer::addNode(int id) {
 	if(!isDynamic) {
 		cerr << "This graph is not dynamic,"
 				"so you must use GraphViewer::addNode(int id, int x, int y) instead.\n"
@@ -109,7 +109,7 @@ bool GraphViewer::addNode(unsigned long id) {
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::addNode(unsigned long id, int x, int y) {
+bool GraphViewer::addNode(int id, int x, int y) {
 	if(isDynamic) {
 		cerr << "This graph is dynamic, "
 				"so the provided x and y values for the node with id "
@@ -122,21 +122,21 @@ bool GraphViewer::addNode(unsigned long id, int x, int y) {
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::addEdge(unsigned long id, unsigned long v1, unsigned long v2, int edgeType) {
+bool GraphViewer::addEdge(int id, int v1, int v2, int edgeType) {
 	char buff[200];
 	sprintf(buff, "addEdge %d %d %d %d\n", id, v1, v2, edgeType);
 	string str(buff);
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::setEdgeLabel(unsigned long k, string label) {
+bool GraphViewer::setEdgeLabel(int k, string label) {
 	char buff[200];
 	sprintf(buff, "setEdgeLabel %d %s\n", k, label.c_str());
 	string str(buff);
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::setVertexLabel(unsigned long k, string label) {
+bool GraphViewer::setVertexLabel(int k, string label) {
 	char buff[200];
 	sprintf(buff, "setVertexLabel %d %s\n", k, label.c_str());
 	string str(buff);
@@ -150,21 +150,21 @@ bool GraphViewer::defineEdgeColor(string color) {
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::removeNode(unsigned long id) {
+bool GraphViewer::removeNode(int id) {
 	char buff[200];
 	sprintf(buff, "removeNode %d\n", id);
 	string str(buff);
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::removeEdge(unsigned long id) {
+bool GraphViewer::removeEdge(int id) {
 	char buff[200];
 	sprintf(buff, "removeEdge %d\n", id);
 	string str(buff);
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::setEdgeColor(unsigned long k, string color) {
+bool GraphViewer::setEdgeColor(int k, string color) {
 	char buff[200];
 	sprintf(buff, "setEdgeColor %d %s\n", k, color.c_str());
 	string str(buff);
@@ -178,7 +178,7 @@ bool GraphViewer::defineEdgeDashed(bool dashed) {
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::setEdgeDashed(unsigned long k, bool dashed) {
+bool GraphViewer::setEdgeDashed(int k, bool dashed) {
 	char buff[200];
 	sprintf(buff, "setEdgeDashed %d %s\n", k, dashed? "true" : "false");
 	string str(buff);
@@ -192,7 +192,7 @@ bool GraphViewer::defineEdgeCurved(bool curved) {
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::setEdgeThickness(unsigned long k, int thickness) {
+bool GraphViewer::setEdgeThickness(int k, int thickness) {
 	char buff[200];
 	sprintf(buff, "setEdgeThickness %d %d\n", k, thickness);
 	string str(buff);
@@ -206,7 +206,7 @@ bool GraphViewer::defineVertexColor(string color) {
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::setVertexColor(unsigned long k, string color) {
+bool GraphViewer::setVertexColor(int k, string color) {
 	char buff[200];
 	sprintf(buff, "setVertexColor %d %s\n", k, color.c_str());
 	string str(buff);
@@ -220,7 +220,7 @@ bool GraphViewer::defineVertexIcon(string filepath) {
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::setVertexIcon(unsigned long k, string filepath) {
+bool GraphViewer::setVertexIcon(int k, string filepath) {
 	char buff[200];
 	sprintf(buff, "setVertexIcon %d %s\n", k, filepath.c_str());
 	string str(buff);
@@ -234,7 +234,7 @@ bool GraphViewer::defineVertexSize(int size) {
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::setVertexSize(unsigned long k, int size) {
+bool GraphViewer::setVertexSize(int k, int size) {
 	char buff[200];
 	sprintf(buff, "setVertexSize %u %d\n", k, size);
 	string str(buff);
@@ -248,14 +248,14 @@ bool GraphViewer::setBackground(string path) {
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::setEdgeWeight(unsigned long id, int weight) {
+bool GraphViewer::setEdgeWeight(int id, int weight) {
 	char buff[200];
 	sprintf(buff, "setEdgeWeight %u %d\n", id, weight);
 	string str(buff);
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::setEdgeFlow(unsigned long id, int flow) {
+bool GraphViewer::setEdgeFlow(int id, int flow) {
 	char buff[200];
 	sprintf(buff, "setEdgeFlow %u %d\n", id, flow);
 	string str(buff);
